Moved AuthClientAuthConnector to its own file
